Move contributor to separate structure in tests
