+++ tests/test_dialects/test_oep/test_compiler.py
@@ -1,5 +1,5 @@
 from metadata_tool.dialects.oep.compiler import JSONCompiler
-from .internal_structures import metadata_v_1_4
+from ..internal_structures import metadata_v_1_4
 import json
